Phase D1d: PyTorch inference float32 dtype enforcement
Implemented float32 dtype enforcement in PyTorch inference paths to resolve
RuntimeError 'Input type (double) and bias type (float)' errors.

Implementation:
- ptycho_torch/workflows/components.py:443-444: Cast infer_X and infer_coords
  to float32 in _build_inference_dataloader before TensorDataset construction
- ptycho_torch/workflows/components.py:700: Defensive float32 cast before
  Lightning forward pass in _reassemble_cdi_image_torch
- ptycho_torch/inference.py:494-495: Cast CLI data loading to float32/complex64

Tests:
- tests/torch/test_workflows_components.py: Added TestReassembleCdiImageTorchFloat32
  with 2 tests validating float32 preservation and float64→float32 casting
- Targeted tests: 2/2 passing (3.67s)
- Full suite: 233 passed, 16 skipped, 1 xfailed, 1 failed (ZERO new failures)

Integration test advancement:
- test_pytorch_train_save_load_infer_cycle now progresses past dtype error
- New failure is shape mismatch (572 vs 1080) at model.py:366, tracked separately

Artifacts:
- plans/active/INTEGRATE-PYTORCH-001/reports/2025-10-19T110500Z/phase_d2_completion/
  - pytest_dtype_red.log (RED phase with float64 input)
  - pytest_dtype_green.log (GREEN phase, 2 passed)
  - pytest_integration_dtype_green.log (integration test showing dtype fix)

Refs:
- specs/data_contracts.md §1 (float32 requirement)
- plans/active/INTEGRATE-PYTORCH-001/phase_d2_completion.md D1d
- docs/fix_plan.md INTEGRATE-PYTORCH-001-STUBS Attempt #37

+++ ptycho_torch/inference.py
@@ -487,8 +487,12 @@ Examples:
 
     try:
         # Extract data from NPZ
-        diffraction = torch.from_numpy(test_data['diffraction']).to(args.device)
-        probe = torch.from_numpy(test_data['probeGuess']).to(args.device)
+        # DTYPE ENFORCEMENT (Phase D1d): Cast to float32 to prevent Conv2d dtype mismatch
+        # Requirement: specs/data_contracts.md §1 mandates diffraction arrays be float32
+        # Root cause: torch.from_numpy preserves dtype; legacy NPZ files may contain float64
+        # Symptom: RuntimeError "Input type (double) and bias type (float)" in model forward
+        diffraction = torch.from_numpy(test_data['diffraction']).to(args.device, dtype=torch.float32)
+        probe = torch.from_numpy(test_data['probeGuess']).to(args.device, dtype=torch.complex64)
 
         # Handle different diffraction shapes (H, W, n) vs (n, H, W)
         # Per specs/data_contracts.md, diffraction should be (H, W, n)
